micro: split status section components
